backend/requirements.txt
backend/requirements.txt:

flask==3.0.2
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
google-genai==1.63.0
python-dotenv==1.0.1